Add files via upload
Understood the problem statement, which is that the online learning portal X Education sells courses to industry professionals. There are different sources from where we get the potential customers which are called the Leads. Hence we have data of such customers who were earlier leads and were either converted into customers or did not chose to opt for any course.

Our work is to study the profile, education, website metrics etc these features and generate a score for all the Leads showing who are more likely or less likely to opt for the course.